[Semi-Major Update]
Front-end:
- New 'more' modal to display settings & more [wip]
- Ability to toggle leaderboard & chat
- Major improvements to responsiveness
- Swapped leaderboard & chat positions
- Player count added to chat

Backend
- Boilerplate code for new SyncCode system

SyncCode is my solution to persistence and account recovery without the need for users creating accounts.

Pet Daisy is meant to be a hyper casual game where people can hop on and click my friend's dog while having a little chat when they're bored at school. For this reason the account creation and persistence process must be as frictionless as possible. Now, even if cookies are deleted the user has a chance to recover their account. Yay